+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/commands.py
@@ -3,7 +3,6 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.cli.core.util import empty_on_404
 from ._client_factory import cf_container_groups, cf_container
 from ._format import transform_container_group_list, transform_container_group
 
@@ -13,8 +12,7 @@ def load_command_table(self, _):
         g.custom_command('list', 'list_containers', table_transformer=transform_container_group_list)
         g.custom_command('create', 'create_container', supports_no_wait=True,
                          table_transformer=transform_container_group)
-        g.custom_command('show', 'get_container', exception_handler=empty_on_404,
-                         table_transformer=transform_container_group)
+        g.custom_show_command('show', 'get_container', table_transformer=transform_container_group)
         g.custom_command('delete', 'delete_container', confirmation=True)
         g.custom_command('logs', 'container_logs', client_factory=cf_container)
         g.custom_command('exec', 'container_exec')