+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,8 +3,9 @@
 Release History
 ===============
 
-2.1.6
+2.2.0
 +++++
+* BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.
 * `network nic create/update/delete`: Add `--no-wait` support.
 * Added `network nic wait`.
 * `network vnet subnet list`: Argument `--ids` is deprecated.